+++ library/Rules/Length.php
@@ -87,7 +87,7 @@ final class Length extends AbstractRule
     private function extractLength($input): ?int
     {
         if (is_string($input)) {
-            return mb_strlen($input, mb_detect_encoding($input));
+            return (int) mb_strlen($input, mb_detect_encoding($input));
         }
 
         if (is_array($input) || $input instanceof CountableInterface) {
